requirements.txt
amqp==5.2.0
asgiref==3.7.2
async-timeout==4.0.3
billiard==4.2.0
celery==5.3.6
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
Django==5.0.2
django-filter==23.5
djangorestframework==3.14.0
flower==2.0.1
humanize==4.9.0
kombu==5.3.5
Markdown==3.5.2
prometheus-client==0.19.0
prompt-toolkit==3.0.43
python-dateutil==2.8.2
pytz==2024.1
redis==5.0.1
six==1.16.0
sqlparse==0.4.4
tornado==6.4
tzdata==2023.4
vine==5.1.0
wcwidth==0.2.13
